Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  61c6e63a99631e5c7a792d0a9786dc8aa7e27875
Don't redirects anons to /createSession

/createSession is still a significant amount of traffic from enwiki.
We originally redirected the user to /createSession on the local wiki
so we could set a cookie on the local wiki's domain to avoid rechecking
the central wiki. But since we're doing this from javascript now, which
has the localwiki's SOP, we don't need to redirect back.

Bug: 54195
Change-Id: I4e0e7d54ffc1d84de9c3617c887d053f2575194f