Updated mediawiki/extensions Project: mediawiki/extensions/Flow  ec94e4600e65e997a43ba4023fcc477452ce66ce
Get rid of unused file

IIRC, we quickly slapped this together back when header did not have the JS
interaction (like posts had), and we were always sent to the separate
&action=edit-header page, but still wanted to add an editor.

Now, editor & the submissions of data, handling of errors, ... is all JS-based
for header too. The only ones on &action=edit-header will be non-JS users, so it
no longer makes sense to have this file that only loads & unloads the editor on
there.

In a previous patch (I92a4a75093500bf7d50ddfd00e1233676ab23548), I apparently
had missed the reference in Resources.php

Change-Id: I1f6574133bf4f66033cbf647304ed88920fdb743