Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  efcf1fcb5c828ae7404ce404359f5a2e1d71f466
Toolbar edit group widget optimization

The change applies some logic and performance improvement to the toolbar editgroup widget.
The edit group buttons' characteristics are moved to the options object. Except for the
edit button which is displayed initially, the other buttons will be created not before
they are needed for interaction. Additionally, the event handlers on each button have
been merged to a single event handler applied to the edit group node itself.

Change-Id: I80a5f17c24b20c230062630193b3d3360de77697