Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  534fe708c2460b0eba81c8ef9a72b656e62a3a9a
Index and search file text from pdf/djvu files

Requires I22f5fc47 from core to work.

File text is searched and presented to the user if there aren't page text
matches.  We don't help the user jump to the appropriate portion of the
file, though.

DEPLOYMENT: This requires an inplace reindex and then a full reindex of
at least the file namespace to work.

Bug: 6421
Change-Id: I43aefedebada08a75a2e8478afdd1cecb136d5b5