Updated mediawiki/extensions Project: mediawiki/extensions/CreditsSource  fede87dad4f486d5cf61c89ae326022f04915692
Don't pass the Title object as third argument to CreditsSourceAction::getCredits()

Fix for I102e9dcdc (54c5593).

getCredits() is also called from SkinTemplate when showing the credits in the page's
bottom. Instead make getCredits() non-static so that it can call $this->getTitle().

And while I'm at it: remove usage of other global variables: $wgLang and $wgUser
(the latter being implicit).

Change-Id: Ie10606be7f7ea706f398a6d748509a8da2a01f16