Updated mediawiki/extensions Project: mediawiki/extensions/MassMessage  c7538562e6d6ace1bc033f01c019e6b66054f9b0
Store title as a job parameter

In some cases, a namespace on a target site will be an
interwiki prefix on the submission site. This, along with
bug 58524 will cause the namespace to be dropped, so the
job would use the wrong title.

We work around this by storing the real title as text in the job's
parameters. When the job is pulled from the queue on the target
site, the full title, as the user intended, will be evaluated in
the context of the target site.

Bug: 57464
Change-Id: Ie3ecb22001b4f9bc8d9b891cd096061da5151505