Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  5b5c3e74798992eed868712c292fd14c4b8cf230
Fix chunk size on in place reindex

Some old Elasticsearch documentation says that the size parameter is per
shard when scrolling but that is not the case - it is across all shards.
The official Elasticsearch documentation no longer mentions this, so I
assume it has changed.

We used to have a case were we'd send a non-integer size parameter and that
was breaking reindex.  This prevents that as well.

Bug:  58550
Change-Id: I45290b9eb0d25d5ddeb6cebd42631ccfa53175dd