Updated mediawiki/extensions Project: mediawiki/extensions/Flow  0a35773b31b3499581762dab992c30c8a55e1247
Fix post-edit page refreshing.

Previously, initialisation was not being run on the new post, which
Benny tried to fix in change 95048, but introduced bug 57941 in the
process. The correct thing to do is to trigger the flow_init event
on the post container.

The post container was also being put inside the existing container.
This patch changes that to replace the existing container with the
new one from the API.

Bug: 57941
Bug: 56757
Change-Id: I6b2924b6b67b84cedc20a87ad419a44c6b363131
Reverts: 4238cf6e4da9e52f08c9ed954662115df143713e