Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  5597cef6f3496e9e2ae1bbefaf06ce7d757b47cd
Remove some of the permissions so they can be configured per-wiki

Previously this auto-added 'gwtoolset' right to admins, and made
it so admins could add/remove the gwtoolset group.

Instead, only make the gwtoolset group. Let the configuration for
if additional groups get the right, and which groups can add that
right must be specified in the individual wikis config.

After this change is merged, the configuration on commons will
have to be changed.

Bug: 58603
Bug: 58607
Change-Id: I4d9097e59dd91a104f1e3ac4c1dd543d3a37c9f9