Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  644fb984101f953fa987a8d11c1d6e63acc8fcad
Switch index requests to updates with docAsUpsert

Switch all index statements to updates with docAsUpsert to make sure we
don't throw out any fields maintained outside the normal index mechanism
like local_sites_with_dupe.

This should be safe because the merging mechanism that docAsUpsert uses
sits well with us because it won't try to merge lists.  Wonderful!  This
does cost some performance on the Elasticsearch side but we're not really
hurting there.

Bug: 58508

Change-Id: I50c6b501dd7fe568e5e281fbf6ec9f1f4b891247