Updated mediawiki/extensions Project: mediawiki/extensions/JsonConfig  0798e864330d9c4e1b3e619926afdd726cb27db7
Initial implementation of the JSON configuration store.

This extension allows other extensions to store configuration information
as JSON blobs on the wiki. For more details, see RFC at
https://www.mediawiki.org/wiki/Requests_for_comment/Json_Config_pages_in_wiki

The code is partially ported from the Zero extension with a significant
rewrite of various pieces. From the begining it is being structured to
allow seamless transition for the current Zero and Schema namespaces.

The extension is coming along nicely, but badly needs your review...

Change-Id: I1af6ddabc126058cf732318066e0eca4a75cc964