Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  7cd8f3c69ef7bf2fbc140eecc359f33608b3af7c
QA: Add a test that does not run in the build yet.

Merge this in order to pick up the fix from the Nearby
work that causes tests with more than one scenario
to have no profile past the first scenario, and
prevents launching Chrome

This test has no target environment, no steps file yet,
and no pages created.  It is a no-op, but without the
Nearby fix it is difficult to run locally

Change-Id: Ie87ed90fa526fb7e98a5bd37ed6e8d549c26e2ed