Updated mediawiki/extensions Project: mediawiki/extensions/Comments  dcaf5f13c5449fd31965f7c77e16f894ce196c5c
Reduce comment cache time

I (we) have found that having a 24 hour cache for NUMBEROFCOMMENTSPAGE can be
too much, as articles are most frequently commented on in the few hours
after their publishing, and so you can end up with very out of date
figures with a 24 hour cache. I've reduced it to 1 hour here, which should
have an almost insignificant effect in performance, while giving much more
up to date stats.

Change-Id: I9e8bfe1fda0f2c5acc5c2a94b5132af598aeb298