Updated mediawiki/extensions Project: mediawiki/extensions/Flow  d48e96f5d1a8f39a6aec35e10c20fbb239a418c8
Post content with more than 5 lines is showing a scrollbar

It turns out this is caused by the tipsy link, somehow the icon is
creating an overflow even the height/width are the same as icon
height/width.  Adding overflow: hidden seems to fixe the
problem.  Not sure if this is a good fix but this does address
the issue, :)

Change-Id: Iae275a3d1c18e7b9693811e20cf3e742a48f23be