Updated mediawiki/extensions Project: mediawiki/extensions/Flow  21887747b002a1cdb904b897f2a37ff9123a63ca
[Untested] Drop sqlite-specific script to add *_user_ip columns

I'm pretty sure sqlite is capable of ALTER TABLE x ADD y, and some
UPDATE-statements, which is all that patch-remove_usernames.sql does.

I suggest removing this sqlite-specific patch, because it differs
from the default patch: in sqlite, *_user_text columns will be
dropped, in the default patch, these will not be dropped.

Change-Id: Ieae18fa24de8da55c5e1e45e3532a6c2ca53b482