Updated mediawiki/extensions Project: mediawiki/extensions/Thanks  75c88532653ed7c9696314a4731490b5dbba3c56
Remove misleading Thank button from some diffs

To prevent thanking someone for the wrong revision, don't show the Thank
button unless the old revision is the parent of the revision being thanked
for. This needs I6ee956c5 in core to function, but it degrades gracefully
if it's not present.

Bug: 57721
Change-Id: I8f973a8a54db35838ce4c697c0310f67b2dba1a0