Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  38d7c935be8ee63f23e3dc6634bfe59954c4ef5e
Fix alignment issue of ULS trigger

The calculated font-height is about 13.5px on my machine. For some
reason the usage of Autonym font sways the rounding of the container
height to 14px instead of 13px as for other items. I found out that
setting line-height to 1em or 100% makes the trigger size also round
to 13px, but I can't explain why. I didn't see any change in Modern
or Monobook skins (Monobook is broken already with bold and icon
missing).

Bug: 59239
Change-Id: I71ed556f2ef707b7e5f5891d6fc0c88c01f6b0ca