Updated mediawiki/extensions Project: mediawiki/extensions/VectorBeta  9c81f80447df8735b3937ecde4b54b45adcff94c
Removing italic override for blockquotes

Forcing all blockquote text to be italic causes loss of information.
For example, in the attribution it is no longer clear what is the
author and what is the source.

Compare https://www.mediawiki.org/wiki/User:Kaldari/Sandbox with and
without the typography refresh. With the typography refresh activated,
all distinction between italicized words and non-italicized words is lost
(since everything is forced to be italics).

Change-Id: I8aa2a5fd8a893e92cfee3e193d03458b6e06d342