Updated mediawiki/extensions Project: mediawiki/extensions/UploadWizard  25d1e0b31fce7b9694a1cc9ee413bf15b1c501c0
Solves visibility issue on the progress bar

At first it seemed odd to me that the progress bar
appears only on the 2nd upload of a multi-file upload.
But the code seems to suggest that's intentional.
At any rate, this commit should fix the fact that the
progress bar doesn't appear when the 2nd upload starts.

Bug: 51787
Change-Id: I90509045349c46dde8ff1a1b1258b61867c47d55