Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaIncubator  09745eb2d387d1e5e016b5500f4649a7c3faaf75
Use namespace ID for Module: namespace
This commit changes all the proper names for namespaces into their
IDs. This change was meant for the Module: and Module_talk:
namespaces (ID 828 and 829 respectivily) specifically, but for
consistancy in the array, I changed all the values, adding a
comment for easier reading (not that code is easily readable.

This change was necessary as the Scribunto extenion (creates the
Module: namespace) is loaded after the WikimediaIncubator extension
(this extension)

This should squash bug 59187 hopefully

Bug: 59187
Change-Id: If0f030674c728d39876a251a7c35613b6f88d2bf