Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  083b22e169b95b2a2dda9dfc184379620f99a3f2
image url not evaluated

david ran into an issue on commons where the extension was not able to properly
analyze a media-file url; there is no issue on the beta cluster.

the error message currently gives only the url to the media file. this patch adds
any mw http request message that may be available in order to help further
troubleshoot the issue.

also added a $wgCopyUploadProxy test to make sure the proxy option is set for the
HTTP Request when a proxy is being used.

Change-Id: Iac8d72a25b88b9a62d7b95733391a3a6103d0807