Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  8815b53223b5666e27dbf96a4a3f181d503384ae
Fix logic error that caused dual logging

DocumentMissingExceptions aren't really failures as we'd like to
one day actually reattempt them. With the backwards check here,
we're both putting these in the general debug log (where they
belong for now until we fix it), but never in the actual failure
log.

Change-Id: Ibfd28b3bf5c17e6dd3d4a672328702f868c665c2