Updated mediawiki/extensions Project: mediawiki/extensions/Flow  cdc9d4427611012867c84dd5274a42a404f9730f
Display "This comment was <moderated> by <User>" instead of content

This is a left-over from when Templating::getContent accepted a $user param
(which, in this case, was set to null to get the moderated message for
unprivileged users)
We have now a specific method to get just that moderated message.

Also: getModeratedContent has no $format argument.

Change-Id: I46f2adc80698e50986b47246afe114b42ab0a80d