Updated mediawiki/extensions Project: mediawiki/extensions/GuidedTour  b50f89205c1d2ba199351012060e564352121206
Simplify the firstedit tour by removing special VE handling

This was so it would refer to 'edit source' if VE were installed.
However, VE now only modifies the tab text if it's enabled for the
current user.

We can assume that VE is always available in the main namespace
(the only namespace firstedit supports), if it's installed at all.

firsteditve should be used instead if VE is enabled for the user.
For that tour, the text of the wikitext edit tab is irrelevant.

* Move tours from ResourceLoaderRegisterModules to GuidedTour.php
* Remove unneeded messages
* Arrange messages with core first, other extensions next, and GuidedTour at the end

Change-Id: I6b0f60edb7d33ef11ce74518f26d4456a201c57a