Updated mediawiki/extensions Project: mediawiki/extensions/Translate  15bdfe3f802844b104599aafcd737a8f5f5a8922
Don't try to sandbox unsandboxable users

If a user is already past the sandbox stage, redirect them to the
normal userrights process: they can't do anything on the main page.

jgonera> Nikerabbit, I created a new account on translatewiki so that it
has the same user name as on other wikis. I tried translating a message
but I don't have the permissions. Unfortunately, "Get permission" link doesn't
work. It leads to https://translatewiki.net/wiki/Special:TranslationStash
which in turn redirects me to the main page...

Change-Id: I475d99ee47c3307f1abe57f28f1d76067a15a7f8