Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  5ca0d5d4005c3139a41bf49d5d37c20ebc7e0acc
Compacting the interlanguage links with the ULS

Displays a shorter version of the language list with the languages that are more relevant to you.
More info at https://www.mediawiki.org/wiki/Universal_Language_Selector/Design/Interlanguage_links

Due to a problem with duplicate patchsets we had to create a new one. 
The history and info on code and security revisions is available at
https://gerrit.wikimedia.org/r/#/c/104793/

Change-Id: I109bfb10f6181b6408aab170db5e8ca32493d657