Updated mediawiki/extensions Project: mediawiki/extensions/WikiForum  afba9fc49272278a180611886b9b95d55ef0763b
Use placeholder instead of value

For some reason, the placeholder text was being placed in a value
attribute, not a placeholder one as it should be, and as it is in other
areas of the wikiforum code. Using placeholder means the user no longer
has to delete the current text before typing their own. (I have only found
this affecting creating threads, not forums or categories though)

Change-Id: I7e1cafb946871dcd3a848e4106d44f095c92b04b