Updated mediawiki/extensions Project: mediawiki/extensions/VectorBeta  4ef490776e2ac5486fefbe0e43ac9392b941d15c
Typography updates from typography refresh meeting

* Clean up blockquote display
* Increase line-height for content
* TOC redesign
* hatnote redesign
* Remove thumbnail borders
* Tweak thumbnail captions
* Optimize header font display for Linux:
After doing some research and consulting with Linux Wikipedia users,
it was determined that the way to get the best font rendering in
Linux (i.e. the closest to that desired by the designers) was to
utilize the font fallback mappings in Linux, rather than trying to
specify specific free fonts. This change was the consensus decision
of the refresh meeting, but may be revisited in the future.
* Set article title to 1.8em

Note that infobox changes are not part of this changeset, but will
be in a follow-up change.

Change-Id: I8259c92c6b71ea4be04250ad53b248e2efe7870e