Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  dffff9fa5701d5f837f334d781247b3b6f749ea9
Reindex is ok seeing same id twice

I believe the scan process used in reindex can see the same id twice if
the document with that id is updated in the old index.  I doubt it always
sees it, so this isn't something we can rely on.  It is more something we
have to work around.

We work around this by leaving the bulk operation type during the process
set to "index" rather than "create".  "create" might have got us some
performance benefits but was more brittle so it isn't worth it.

Bug: 60854
Change-Id: I4d1197f6a5df19e4435dc00cc341254c80149f46