Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  0668ee9448ced9e8380d808c9f3ecff83b4e1ae5
Use redirects in prefix and go searches

It looks like the old search used to include redirects in the prefix search
and Cirrus didn't.  So this makes it do that.  The redirects show up as
though they were a regular title in the prefix search select them or just
hit enter in the search box it'll bounce through the redirect to the target.

DEPLOYMENT: Doesn't do anything until an in place reindex.  Shouldn't
hurt anything before that.

Bug: 60959
Change-Id: I924200f86d8242b53f8b5d67e40a4f09ca725722