Updated mediawiki/extensions Project: mediawiki/extensions/ThrottleOverride  120386dfde326a99c5fb0a4c5e5c83cf108f4147
Fixed filtering	for target field on Special:ThrottleOverride

Unfortunately, HTMLForm's filtering feature is pretty terrible
and does not work as expected. This patch removes filtering on
the target field in Special:ThrottleOverride and does the filtering
manually in the	 onSubmit() function.

This also simultaneously adds a variable and check to make sure
range exemptions are not too large.

Bug: 60409
Change-Id: I9f2d4a5830ddcf69fecac87900412fb55b21ee7b