Updated mediawiki/extensions Project: mediawiki/extensions/UniversalLanguageSelector  1c3a935ef0fd01c38b9e1ad9c5765f411c632d8c
Try to restore the old preferences, if any, if possible

Preferences were saved using jStorage. The browsers which ULS supports now
have wide native support for local storage, so we removed jquery.jStorage
in I09b2bd52972.

The old local storage key was 'jStorage' as put by jquery.jStorage. We now
check if that key is present. If so we migrate the preferences and remove
the old key to save space. We are not aware of anything else using this
module.

Change-Id: I995a32c4cb7a3bc818fe255f92dbc972211cb112