Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  d71447266230acfb7b5f0cd72039dfb7ede6845e
Make MVEL way faster

Using ".empty" in MVEL is slow when fields are actually empty.  Instead,
use ".isEmpty()".  Against enwiki I measure an order of magnitude performance
improvement on queries that hit lots and lots of results and can't push
the script scoring in the rescore.  3 seconds down to half a second.

Change-Id: Idd394ddc9477fced9130c0aab6f787eb2510f6dc