Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  aeb6d6bc0887839086a8f8f02bd4040fabba996d
Add support for "%{fieldName}i"-style format specifiers

We need to add a user-agent field matcher to the cilent-side event parser's
format string, but we don't have a specifier for it. We could pick an
unassigned letter and turn it into a user-agent format specifier (e.g. '%a').
But that's ad-hoc. It's nicer to support varnishncsa's notion of naming
capturing groups via squiggly brackets.

Thus, introduce the following format specifiers:

* "%{..}i": Tab-delimited string
* "%{..}s": Space-delimited string
* "%{..}d": Integer

..where '..' represents the desired name for the capturing group.
With this change, we can deprecate %l and %n, so do that too.

As an example, we can replace this format string:

    %q %l %n %t %h

With this one:

    %q %{recvFrom}s %{seqId}d %t %h %{userAgent}i

Change-Id: I07c6c785648397c2fe51247edb27653885c0cbed