Updated mediawiki/extensions Project: mediawiki/extensions/EducationProgram  d60a17cb0fa3c1581684efbdc0d3106fbd2b8d80
Remove the "field" and "level" options in the course edit view

Remove "field" (of study) and "level" (of students) from the form
for editing a course. The data for those fields aren't displayed
anywhere, and they are optional, so they don't serve a purpose.
Removing them from the edit interface but leaving the tables alone
will be minimally invasive.

Change-Id: I599044687ea44fb796e644f4da10326728571f78