Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5e9a3c340e6090c177c40acf0f7c93b1bad4b279
Fix search result display for title matches

CirrusSearch wraps title matches in a span tag, which
does not work with our LinkBegin hook handler.

The span tag is normally used to provide bold highlighting
of matches.

Perhaps in the future, we can improve the entity link formatting
to provide highlighting of both Q## titles and labels.

Bug: 62348
Change-Id: I1a4e9cb622d00c72177a415c39ac627c88536b0a