Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  d1d608c33a007c731deb9528ff256c4c8e0656ba
Quieting down the error logs
This statement is creating a clog on the user's first request,
when no card type has yet been selected. Which is normal, and not
an error. We only want to know about this when something has been
selected that we don't have in the map.

Change-Id: I0f3c05583b4ac27fc26e3d74943e5ae8972fb901