Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  fa60e61ad30a4325804a8859d9f22904af931fa9
Found and altered the one place in the code that might explain
these infrequent and weird loglines I'm seeing sometimes in
the payments.error log (in production).
If this changes things: Transactions that hit this block get stuck
until they are manually unstuck. Teach orphans.php how to handle
them for real.

Change-Id: I69b252529aaa8ee075351c049fea9abd59e9174e