Updated mediawiki/extensions Project: mediawiki/extensions/DonationInterface  b27d1eaf1b4b8b6aa008d37727938f7d27bec177
WIP: Bigger stub for worldpay

It can now make a transaction! and render a form!

There are two things that still need to be injected into the form:
* The one time token 'one_time_token' which is an identifier the
  tokenizer needs in order to associate a user CC data with us
* 'wp_process_url' which is the URL we need to submit tokenization
  requests to.

Additionally; though this patch makes available a worldpay.js module,
it doesn't yet do anything with it. This script should submit the tokenization
request as a AJAX request; get the response, clear the form of incriminating
fields, and then submit a payment to us (via the donate JS API? or via form
submission.)

Change-Id: I889748f9709be8cf7070bb083864e304572942cd