Updated mediawiki/extensions Project: mediawiki/extensions/Comments  754bbf81ea79b6b305f4ca379d93ca149da88dcd
Added username as parameter to MediaWiki:Comments-reply-to

The interface message  needs to have the username as a parameter
since the subject-object order varies between languages.
Hence, for some languages, like hindi(hi), the username will need
to be before the the text for the message to make sense.
Added gender support to the message.

Bug: 43696
Change-Id: I9cfd71bf7fba134c2d1e22640600bcc6e209e06b