Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a9edd2029e76c0a34a809aa4b730ecebe6affe91
Don't trigger warning if "bot" parameter is missing in API request

I don't think it's required to always have the "bot" parameter set
in the request. It could be "0", empty or simply not exist. All
other parameters are optional, so shouldn't this one too? The
inline documentation doesn't say it's required.

For some strange reason this triggers "Undefined index" errors in a
completely unrelated patch. I didn't spent time on deeper
investigation since I think this is a proper solution.

Change-Id: Ie0650ebe41a328f0576e304c67a7c1f852a0e72b