Updated mediawiki/extensions Project: mediawiki/extensions/BetaFeatures  83f1168f89f26cbaee56917f3748a9c4b3f82d9f
feature-discussion.svg: Snap to one-pixel grid

This makes the SVG look a lot sharper on standard (non-HiDPI) screens,
and probably a bit sharper on the HiDPI ones too (although some points
were already snapped to a half-pixel grid).

The sizes changed a bit (most notably the white divide between the two
bubbles is now 1px wide instead of ~1.5px, and the icon itself is 29px
wide instead of 28.5px), but it looks just as neat to me.

The SVG file was compressed using svgo and optimized further by hand.

Also regenerated the PNG fallback from the new image and recompressed
it using pngcrush.

Change-Id: I832eacb42d8ffdfe3c1cb3afe58d527b7881aef3