Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  04b5c72bd10e40301fe2e91ec4d91ccba3da10a3
QA: this test now works for both FF and Chrome

Also, the upload step now exactly matches the upload step in
the UploadWizard repo, which is known to be correct.

I created Bug 62888 to remove the duplicate code by abstracting
it to the mediawiki_selenium gem where it can be shared among
repos.

See that bug for more information.

Change-Id: I6c421f1eae186f5a30edaba069cb4c3a01fd47a1