Updated mediawiki/extensions Project: mediawiki/extensions/SemanticForms  6d1920c6680c95fd3a7901bf48e2364dfa2f473e
Select all/Select none links for checkboxes input type

This patch will add the parameters 'show select all' and 'hide select all' to the checkboxes input type, where 'show select all' will
attach links to the respective field of checkboxes that allow the selection or deselection of all checkboxes at once and 'hide select all'
will force these links not to be shown.

In the absence of both parameters the links will be shown when the number of checkboxes equals or exeeds the value stored in the newly
introduced global variable $sfgCheckboxesSelectAllMinimum. This variable defaults to 10 and may be adjusted in the LocalSettings.php.

A JavaScript enabled browser is necessary for this to work. JavaScript disabled browsers will not see any difference to the current behavior.

Change-Id: I809d10517ab89e38550340779d4415509d77ee63