Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  9296f5949fa834e33b4e38679e316334d49fda9f
Fix RuntimeException in Special:Newitem with SpamBlacklist

All of the tests seem to miss this again as it only
appears when using the SpamBlackList hook!

This change reverts the change of behaviour introduced
by change Iacd00a3aa3d48b54f8e

Previous to the above change wbEntityId was '' if
there was no entity ID!
Previously we used getFormattedIdForEntity in EntityView
which still returns this '' when there is no entityid!

Bug: 62965
Change-Id: I5779ffc845de9c85418ae760ced03d43543f41c6