Updated mediawiki/extensions Project: mediawiki/extensions/InputBox  7096213874e2faf5141d707866b446d940553984
Adding a JS Class to inputBox to grey out Submit button.

If the user does not enter a title in the <inputbox> for type=create
and hits the submit button, it opens the Main page for editing (or
view-source) due to empty title= param in URL. Greying out the submit
button until user enters the title solves the issue.

Bug: 61606
Bug: 62975
Change-Id: I5c6c34b90c6c499ef7cb96ad8d47fdbbcbbcd8a8