Updated mediawiki/extensions Project: mediawiki/extensions/EventLogging  fb3f66043e3be07a3f2dcc94405dcbe6be42c37f
Allow EventLogging events to be logged via mw#track

This patch registers EventLogging as a subscriber to the 'event.*' namespace of
mw#track. The subscriber function interprets the topic suffix as the schema
name. It then uses mw.loader#using to ensure the relevant schema module is
loaded. Finally, it logs the event.

Thus, usage such as:

    mw.eventLog.logEvent( 'NavigationTiming'. { .. } );

..can be replaced with:

    mw.track( 'event.navigationTiming', { .. } );

Because mw#track is a part of core, it can freely be used in extension code
without introducing a hard dependency on EventLogging. When EventLogging is not
loaded, the event is simply not handled.

The subscriber code is housed in a new module, 'ext.eventLogging.subscriber',
which is loaded by default on every page.

See also related change I2d29315f9.

Change-Id: I0cee7a03e73a26a24e6a94c245b4a47aba7f4688