Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  0ddc162fa08e63aab3fa4f68f1ace7668004b6b0
Make entityselector always trigger the aftersetentity event

This change also makes the EntityIdInput not listen on entityselectorselect, which
is a quite useless event because it happens before the entity actually is set.

In contrast to If62e8ea (which had the same purpose) this doesn't break setting qualifiers
on wikibase-item type claims.

Bug: 63057
Bug: 62868
Change-Id: I61c0f4aee14fe9c791006bd13c82f82ebe5f4f96