Updated mediawiki/extensions Project: mediawiki/extensions/GWToolset  dd0757d58f25b722d34548e00fcfad03c25f43ec
Correct the format of the wikitext output

* allow for empty template parameters
* add =={{int:filedesc}}== to the top of the wiki text
* move gwtoolset parameters out of the artwork template into their own template
* move the categories to the bottom of the wiki text output
* add whitespace between the different sections within the wiki text

another request, that was made that could not be handled here, was to leave
the template parameter order as it is within the actual template; e.g.,
not always alphabetical. this could not be dealt with here because
GWToolset is using templatedata to retrieve the template parameters and
templatedata is alphabetising the parameters. does this need to be dealt
with as a templatedata bug?

@see https://commons.wikimedia.org/w/api.php?action=templatedata&titles=Template:Artwork

Bug: 63168
Change-Id: I4f82c37e8316737c5f8339e2388c1b30143e5204